Remove IAudioRender class
