+++ OMXPlayerAudio.h
@@ -65,7 +65,7 @@ protected:
   std::string               m_device;
   bool                      m_use_passthrough;
   bool                      m_use_hw_decode;
-  IAudioRenderer::EEncoded  m_passthrough;
+  COMXAudio::EEncoded       m_passthrough;
   bool                      m_hw_decode;
   bool                      m_boost_on_downmix;
   bool                      m_bAbort;
@@ -97,7 +97,7 @@ public:
   bool AddPacket(OMXPacket *pkt);
   bool OpenAudioCodec();
   void CloseAudioCodec();      
-  IAudioRenderer::EEncoded IsPassthrough(COMXStreamInfo hints);
+  COMXAudio::EEncoded IsPassthrough(COMXStreamInfo hints);
   bool OpenDecoder();
   bool CloseDecoder();
   double GetDelay();
